Change equipment cost type to integer

+++ core/equipment/__init__.py
@@ -2,7 +2,7 @@ from typing import Protocol
 
 class Equipment(Protocol):
     name: str
-    cost: float
+    cost: int
 
     def __repr__(self) -> str:
         return f'<{self.__class__.__name__}: {self.name!r}>'
